@@ -2,10 +2,10 @@
   "name": "parcel-antd",
   "version": "1.0.0",
   "description": "Use antd in parcel",
-  "main": "index.js",
+  "source": "src/index.html",
   "scripts": {
-    "dev": "parcel ./public/index.html",
-    "build": "parcel build ./public/index.html"
+    "dev": "parcel",
+    "build": "parcel build"
   },
   "keywords": [
     "Ant Design",
@@ -15,20 +15,20 @@
   "author": "afc163<afc163@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.10.2",
-    "@babel/preset-env": "^7.10.2",
-    "@babel/preset-react": "^7.10.1",
-    "@types/node": "^14.0.13",
-    "@types/react": "^16.9.36",
-    "@types/react-dom": "^16.9.8",
-    "babel-plugin-import": "^1.13.0"
+    "@parcel/config-default": "^2.7.0",
+    "@parcel/core": "^2.7.0",
+    "@parcel/resolver-default": "^2.7.0",
+    "@parcel/transformer-less": "2.7.0",
+    "@parcel/validator-typescript": "^2.7.0",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "parcel": "^2.7.0",
+    "process": "^0.11.10",
+    "typescript": "^4.8.2"
   },
   "dependencies": {
-    "antd": "^3.26.18",
-    "less": "^3.11.3",
-    "parcel-bundler": "^1.12.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "typescript": "^3.9.5"
+    "antd": "^4.23.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
