@@ -1,9 +1,14 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
+import React, { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 
 import App from "./App";
+import "./index.less";
 
-ReactDOM.render(
-  <App title="Ant Design & Parcel." />,
-  document.getElementById("root")
+const rootElement = document.getElementById("app") as HTMLElement;
+const root = createRoot(rootElement);
+
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>
 );
